// src/views/detail/IPro.d.ts
export default interface IPro {
  banners?: string[]
  brand?: string
  category?: string
  desc?: string
  discount?: number
  img1?: string
  img2?: string
  img3?: string
  img4?: string
  isrecommend?: number
  issale?: number
  isseckill?: number
  originprice?: number
  proid?: string
  proname?: string
  sales?: number
  stock?: number
}